ATD Team Sheet Bot/requirements.txt
discord.py==2.3.2
gspread==5.12.2
oauth2client==4.1.3
pandas==2.1.4
python-dotenv==1.0.0